Performance improvements - use running-config.xml
Get & Show APIs for all the namespaces use the PanOS XML APIs
to retreive the configuration from the FW device. XML APIs are
pretty slow - their response time is between 300 and 500 ms.
As a result, for clients that request lots of configuration such
as OpenConfig plugin, the user experience is really bad - based
on the configuration, certain GNMI Get operations take up to
35mins.

XML APIs are unnecessary when the pango and the target device
are on the same host. In that case, the configuration can simply
be retrieved from the running-config.xml file. This commit
introduces that capability:

* Allow the client to control whether running-config.xml is
used or not.
* Introduce the API LoadRunningConfig which allows the client
to control when the running-config.xml is read. pango will
read it upon the call to Initialize.
* Get and Show use the xmlquery package to extract the
configuration from the running-config.xml.